aws_lambda_swipeat/DockerLambda/requirements.txt
requests
boto3
pandas
numpy
tensorflow-cpu
pyrebase
PyJWT